[CUDA,OpenCL] Use size_t type for batch size

+++ include/argon2-cuda/kernels.h
@@ -19,7 +19,7 @@ class KernelRunner
 private:
     std::uint32_t type, version;
     std::uint32_t passes, lanes, segmentBlocks;
-    std::uint32_t batchSize;
+    std::size_t batchSize;
     bool bySegment;
     bool precompute;
 
@@ -43,16 +43,16 @@ public:
     std::uint32_t getMinJobsPerBlock() const { return 1; }
     std::uint32_t getMaxJobsPerBlock() const { return batchSize; }
 
-    std::uint32_t getBatchSize() const { return batchSize; }
+    std::size_t getBatchSize() const { return batchSize; }
 
     KernelRunner(std::uint32_t type, std::uint32_t version,
                  std::uint32_t passes, std::uint32_t lanes,
-                 std::uint32_t segmentBlocks, std::uint32_t batchSize,
+                 std::uint32_t segmentBlocks, std::size_t batchSize,
                  bool bySegment, bool precompute);
     ~KernelRunner();
 
-    void writeInputMemory(std::uint32_t jobId, const void *buffer);
-    void readOutputMemory(std::uint32_t jobId, void *buffer);
+    void writeInputMemory(std::size_t jobId, const void *buffer);
+    void readOutputMemory(std::size_t jobId, void *buffer);
 
     void run(std::uint32_t lanesPerBlock, std::uint32_t jobsPerBlock);
     float finish();
